+++ powerauth-data-adapter/src/main/java/io/getlime/security/powerauth/app/dataadapter/repository/SMSAuthorizationRepository.java
@@ -1,0 +1,31 @@
+/*
+ * Copyright 2017 Lime - HighTech Solutions s.r.o.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.getlime.security.powerauth.app.dataadapter.repository;
+
+import io.getlime.security.powerauth.app.dataadapter.repository.model.entity.SMSAuthorizationEntity;
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Component;
+
+
+/**
+ * Crud repository for persistence of SMS OTP authorization messages.
+ *
+ * @author Roman Strobl, roman.strobl@lime-company.eu
+ */
+@Component
+public interface SMSAuthorizationRepository extends CrudRepository<SMSAuthorizationEntity, String> {
+
+}
