--- powerauth-data-adapter/src/main/java/io/getlime/security/powerauth/lib/dataadapter/DataAdapterApplication.java
@@ -1,31 +1,0 @@
-/*
- * Copyright 2017 Lime - HighTech Solutions s.r.o.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package io.getlime.security.powerauth.lib.dataadapter;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-/**
- * Spring Boot application main class.
- * @author Roman Strobl
- */
-@SpringBootApplication
-public class DataAdapterApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(DataAdapterApplication.class, args);
-    }
-}
