--- powerauth-data-adapter/src/main/java/io/getlime/security/powerauth/lib/dataadapter/configuration/WebApplicationConfig.java
@@ -1,73 +1,0 @@
-/*
- * Copyright 2017 Lime - HighTech Solutions s.r.o.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package io.getlime.security.powerauth.lib.dataadapter.configuration;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.util.ISO8601DateFormat;
-import com.fasterxml.jackson.datatype.joda.JodaModule;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.http.converter.HttpMessageConverter;
-import org.springframework.http.converter.json.Jackson2ObjectMapperFactoryBean;
-import org.springframework.http.converter.json.MappingJackson2HttpMessageConverter;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurerAdapter;
-
-import java.util.List;
-
-/**
- * Default Web Application Configuration.
- *
- * @author Roman Strobl, roman.strobl@lime-company.eu
- */
-@Configuration
-public class WebApplicationConfig extends WebMvcConfigurerAdapter {
-
-    /**
-     * Custom object mapper to make sure that dates and other values serialize
-     * correctly.
-     *
-     * @return A new object mapper.
-     */
-    private ObjectMapper objectMapper() {
-        Jackson2ObjectMapperFactoryBean bean = new Jackson2ObjectMapperFactoryBean();
-        bean.setIndentOutput(true);
-        bean.setDateFormat(new ISO8601DateFormat());
-        bean.afterPropertiesSet();
-        ObjectMapper objectMapper = bean.getObject();
-        objectMapper.registerModule(new JodaModule());
-        return objectMapper;
-    }
-
-    /**
-     * Set custom JSON converter.
-     *
-     * @return New custom converter with a correct object mapper.
-     */
-    private MappingJackson2HttpMessageConverter mappingJackson2HttpMessageConverter() {
-        MappingJackson2HttpMessageConverter converter = new MappingJackson2HttpMessageConverter();
-        converter.setObjectMapper(objectMapper());
-        return converter;
-    }
-
-    /**
-     * Register the JSON converters.
-     */
-    @Override
-    public void configureMessageConverters(List<HttpMessageConverter<?>> converters) {
-        converters.add(mappingJackson2HttpMessageConverter());
-        super.configureMessageConverters(converters);
-    }
-
-}
