--- powerauth-data-adapter/src/main/java/io/getlime/security/powerauth/lib/dataadapter/exception/SMSAuthorizationFailedException.java
@@ -1,77 +1,0 @@
-/*
- * Copyright 2017 Lime - HighTech Solutions s.r.o.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package io.getlime.security.powerauth.lib.dataadapter.exception;
-
-/**
- * Exception used for cases when SMS OTP authorization fails.
- *
- * @author Roman Strobl, roman.strobl@lime-company.eu
- */
-public class SMSAuthorizationFailedException extends Exception {
-
-    private Integer remainingAttempts;
-
-    /**
-     * Default constructor.
-     */
-    public SMSAuthorizationFailedException() {
-    }
-
-    /**
-     * Constructor with authorization failure message.
-     *
-     * @param message Authorization failure message.
-     */
-    public SMSAuthorizationFailedException(String message) {
-        super(message);
-    }
-
-    /**
-     * Constructor with authorization failure message and cause.
-     *
-     * @param message Authorization failure message.
-     * @param cause   Cause, original exception.
-     */
-    public SMSAuthorizationFailedException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    /**
-     * Constructor with cause.
-     *
-     * @param cause Cause, original exception.
-     */
-    public SMSAuthorizationFailedException(Throwable cause) {
-        super(cause);
-    }
-
-    /**
-     * Get number of remaining authentication attempts.
-     * @return Get remaining attempts.
-     */
-    public Integer getRemainingAttempts() {
-        return remainingAttempts;
-    }
-
-    /**
-     * Set number of remaining authentication attempts.
-     * @param remainingAttempts Number of remaining attempts.
-     */
-    public void setRemainingAttempts(Integer remainingAttempts) {
-        this.remainingAttempts = remainingAttempts;
-    }
-}
