--- powerauth-data-adapter/src/main/java/io/getlime/security/powerauth/lib/dataadapter/service/DataAdapterI18NService.java
@@ -1,50 +1,0 @@
-/*
- * Copyright 2017 Lime - HighTech Solutions s.r.o.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package io.getlime.security.powerauth.lib.dataadapter.service;
-
-import org.springframework.context.MessageSource;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.support.ReloadableResourceBundleMessageSource;
-import org.springframework.stereotype.Service;
-
-/**
- * Service for i18n of Data Adapter.
- *
- * @author Roman Strobl, roman.strobl@lime-company.eu
- */
-@Service
-public class DataAdapterI18NService {
-
-
-    /**
-     * Data adapter I18N service constructor.
-     */
-    public DataAdapterI18NService() {
-    }
-
-    /**
-     * Get MessageSource with i18n data.
-     *
-     * @return MessageSource.
-     */
-    @Bean
-    public MessageSource messageSource() {
-        ReloadableResourceBundleMessageSource messageSource = new ReloadableResourceBundleMessageSource();
-        messageSource.setBasename("classpath:/static/resources/messages");
-        messageSource.setDefaultEncoding("UTF-8");
-        return messageSource;
-    }
-}
